Start to flesh out ShowActivityRecordsCommand

+++ src/main/java/org/oddys/timetracking/dao/ActivityRecordDao.java
@@ -8,5 +8,5 @@ import java.util.List;
 public interface ActivityRecordDao {
     long getNumberOfRows() throws DaoException;
 
-    List<ActivityRecord> findActivityRecords(int currentPage, int RecordsPerPage) throws DaoException;
+    List<ActivityRecord> findActivityRecords(long currentPage, int RecordsPerPage) throws DaoException;
 }
